improve build

+++ test-run.sh
@@ -1,15 +1,16 @@
 #!/bin/bash
 
-if [ -d ./extensions ]; then
-    rm -Rf ./extensions
+if [[ "$1" == "--checkout" ]]; then
+    if [ -d ./extensions ]; then
+        rm -Rf ./extensions
+    fi
+    mkdir ./extensions
+
+    echo "## Checkout Phase"
+    ./scripts/checkout-modules.sh
+    echo "## Patch Phase"
+    ./scripts/merge-prs.sh
 fi
-mkdir ./extensions
-
-echo "## Checkout Phase"
-./scripts/checkout-modules.sh
-
-echo "## Patch Phase"
-./scripts/merge-prs.sh
 
 echo "## Build Phase"
 ./scripts/build-jdt.sh && \
